Crate greenhorn

Greenhorn - API Documentation

Greenhorn is a rust library for building desktop applications with web technologies in (almost) pure rust.

This is accomplished by separating the application into a server-side process (the backend) and web view implemented in javascript (the frontend). While most HTML-based desktop applications leave state synchronization up to the application logic, this library synchronizes its state at DOM-level. Thus, the user may implement the application logic purely in the backend using rust. This facilitates the integration of a desktop GUI with system services and simplifies application development considerably.

Features

Elm-like architecture but also supports components

Components support fine-grained update/render cycle

Components are owned by the application state and may interact with each other using events

Macros to write SVG and HTML in-line with Rust code

Most tasks can be accomplished using pure-rust. If required, injecting and calling js is supported.

Built-in performance metrics

Spawning system dialogs

This crate does not itself implement a frontend. A frontend is implemented in greenhorn_web_view. It makes use of web_view and tinyfiledialogs-rs.

Example

use greenhorn::prelude::*;
use greenhorn::html;

struct MyApp {
    text: String,
}

enum MyMsg {
    Clicked(DomEvent),
    KeyDown(DomEvent),
}

impl Render for MyApp {
    type Message = MyMsg;

    fn render(&self) -> Node<Self::Message> {
        html!(
            <div #my-app>
                <button type="button"
                        @keydown={MyMsg::KeyDown}
                        @mousedown={MyMsg::Clicked}>
                    {&self.text}
                </>
            </>
        ).into()
    }
}

impl App for MyApp {
    fn update(&mut self,msg: Self::Message,ctx: Context<Self::Message>) -> Updated {
        match msg {
            MyMsg::Clicked(evt) => self.text = "Button clicked!".into(),
            MyMsg::KeyDown(evt) => self.text = "Button keypress!".into()
        }
        Updated::yes()
    }

}

Acknowledgments

The concpet of this library is not at all new. It was already implemented before at least by the Threepenny-GUI library. The API was inspired by the many great rust frontend libraries:

Yew

Seed-rs

Re-exports

pub use crate::component::Component;

pub use crate::component::Updated;

pub use crate::runtime::Runtime;

pub use crate::runtime::RuntimeControl;

pub use crate::websockets::WebSocketPipe;

Modules

any
This module provides the AnyApp wrapper which allows dynamically exchanging components with different message types.

blob
Supports syncing binary data from backend to frontend. Useful for images, media files, ...

component
This module defines components, which support a fine-grained render() cycle.

components
Provides a set of built-in and commonly used components

context
The context module exports the Context type, which is passed into each update() cycle and allows interfacing with system services.

dialog
Allows spawning native dialogs such as file-open dialogs and retrieving their results. This module implements a set of built-in system dialogs which can be shown using Context::dialog().

dom
This module exposes the DomEvent data type which is used to communicate DOM events from the frontend to the backend.

event
This module defines the Event type which facilitates communication between components.

node
Defines Node<T> type for building DOMs in pure rust

node_builder
Rust API for building DOM nodes. Alternative to the html!() and svg!() macros

pipe
This modules defines an interface for the communication interface between frontend and backend.

platform

prelude
Prelude, use greehorn::prelude::* imports the most important symbols for quick access

runtime
Implments the Runtime type, which executes the render/update cycle of the application

service
Supports spawning tasks running on the frontend. Experimental, might be removed in the future. This module exposes the Service trait, which allows writing custom agents communicating with the frontend.

websockets
This modules adds a Pipe implementation based on WebSockets.

Macros

html
Proc macro to generate HTML Nodes implementing a JSX like syntax.

svg
Proc macro to generate SVG Nodes implementing a JSX like syntax.

Structs

Id
Type to produce unique IDs within the process.

Traits

App
Marks a type as render-able as well as update-able

Render
Marks a type as render-able to the DOM.