Module greenhorn::any

This module provides the AnyApp wrapper which allows dynamically exchanging components with different message types.

Motivating Example

Given a property pane which shows the user different properties based on his selection. In a scene, any object could be selected and upon changing the selection, the property view is exchanged.

In this case, the AnyApp may be used to dynamically switch between the property views of different types, without wrapping each possible message type into an enum.

struct PropertyView {
    object: Object,
    view: Option<AnyApp>,
}

impl PropertyView {
    fn show_object_properties(&mut self) {
        self.view = Some(AnyApp::new(self.object.clone()));
    }
}

#[derive(Clone)]
struct Object(Arc<Mutex<ObjectData>>);

struct ObjectData {
    // ...
}

impl Render for Object {
    // ...
}

impl App for Object {
    // ...
}

Structs

AnyApp
Wraps a type implementing App, and as a consequence also Render), and exposes a new App/Render implementation with the dynamic AnyMsg type.

Type Definitions

AnyMsg
Type alias for the dynamic message type used by AnyApp.