Struct greenhorn::blob::Blob

pub struct Blob { /* fields omitted */ }

Blobs allow transferring binary data from backend to frontend.

Blob objects have an assoociated id and hash fields to make them comparable and identifiable. The user of this type is responsible for providing unique id and hash value combinations.

The id field is supposed to identify the purpose of the Blob within the application. For example, it might be used to point to an image in the frontend. The hash field should reflect the content of the Blob. It might also be considered the version of the underlying data.

A MIME-type may be associated with a Blob object. For example, this is allows providing meta information for media files or similar.

Blobs require manual javascript interaction on the frontend side. Two hooks are provided:

on_change(): Called whenever the hash of a Blob changes, but the id remains the same

on_add(): Called when at least the id changes

Blobs are created using BlobBuilder objects, which are obtained by calling Blob::build().

Example

struct MyImage {
    blob: Blob,
}

impl MyImage {
    fn new() -> Self {
        // hash can be constant, if we don't plan to change the content
        let blob = Blob::build(0)
            .mime_type("image/png")
            .data(vec![1,2,3])      // add image data here!
            .build();
        MyImage { blob }
    }
}

impl Render for MyImage {
    type Message = ();

    fn render(&self) -> Node<Self::Message> {
        let js = format!("{{
            var blob = app.getBlob({});
            var img_url = URL.createObjectURL(blob.blob);
            event.target.src = img_url;
        }}", self.blob.id().data());
        Node::html().elem("img")
            .js_event("render", js)
            .add(&self.blob)
            .build()
    }
}

Implementations

impl Blob

pub fn build(hash: u64) -> BlobBuilder

Start building a Blob object.

pub fn id(&self) -> Id

Returns the id associated with this blob.

Note that the id and the hash define a unique key to the Blob. The id should identifies the blob within the applications data-flow and is usually auto-generated.

pub fn hash(&self) -> u64

Returns the hash of the data associated with this blob.

Note that the id and the hash define a unique key to the Blob. The hash should represent the version of the data underlying the Blob. Thus, if the hash changes, the Blob is considered changed and re-transmitted to the frontend.

pub fn data(&self) -> &Vec<u8>

Returns a reference to the data underlying this Blob

pub fn mime_type(&self) -> &str

Returns the mime-type of the data

pub fn on_add(&self) -> Option<&str>

Returns the registered javascript function which is called when the Blob is added to the frontend.

pub fn on_change(&self) -> Option<&str>

Returns the registered javascript function which is called if the content of the Blob changes, i.e. if the hash of the Blob changes but the Id remains the same.

Example

fn render_blob() -> Blob {
    let blob_id = Id::new();
    let js = format!("{{
        var elem = document.getElementById('my-html-id');
        var blob = app.getBlob({});
        var img_url = URL.createObjectURL(blob.blob);
        elem.src = img_url;
    }}", blob_id.data());

    Blob::build(0)
        .data(vec![1,2,3])
        .mime_type("image/png")
        .id(blob_id)
        .on_change(js)
        .into()
}