Module greenhorn::component

This module defines components, which support a fine-grained render() cycle.

Types which are Render or App can we wrapped into a Component<_> to support fine-grained render() cycles. During the update() cycle components inform the runtime (using Updated) whether re-rendering of their associated DOMs is required. The runtime tracks all components and, if required, calls their render() functions. This reduces the size of the DOM to be diffed.

Note that Component<_> instances also require their contents to be Send to support parallel rendering.

Example

// --------- This would be typically in its own module or even crate ------------

struct Button {
    text: String,
    clicked: Event<()>,
}

impl Button {
    fn new(text: &str) -> Self {
        Self { text: text.to_string(), clicked: Event::new() }
    }
}

enum ButtonMessage {
    MouseDown(DomEvent),
    KeyDown(DomEvent),
}

impl Render for Button {
    type Message = ButtonMessage;

    fn render(&self) -> Node<Self::Message> {
        html!(
            <button type="button"
                    @keydown={ButtonMessage::KeyDown}
                    @mousedown={ButtonMessage::MouseDown}>
                {&self.text}
            </>
        ).into()
    }
}

impl App for Button {
    fn update(&mut self, msg: Self::Message, ctx: Context<Self::Message>) -> Updated {
        match msg {
            ButtonMessage::MouseDown(evt) => {
                if evt.into_mouse().unwrap().button == 1 {
                    ctx.emit(&self.clicked, ());
                }
            },
            ButtonMessage::KeyDown(evt) => {
                if evt.into_keyboard().unwrap().key == "Enter" {
                    ctx.emit(&self.clicked, ());
                }
            }
        }
        Updated::no()
    }
}

// --------- Probably in a different file such as app.rs ------------

struct MyApp {
    ok_button: Component<Button>,
    cancel_button: Component<Button>,
}

impl MyApp {
    fn new() -> Self {
        Self {
            ok_button: Component::new(Button::new("OK")),
            cancel_button: Component::new(Button::new("Cancel")),
        }
    }
}

enum MyMsg {
    OkButton(ButtonMessage),
    CancelButton(ButtonMessage),
    OkClicked,
    CancelClicked,
}

impl Render for MyApp {
    type Message = MyMsg;

    fn render(&self) -> Node<Self::Message> {
        html!(
            <div #my-buttons>
                {self.ok_button.mount().map(MyMsg::OkButton)}
                {self.ok_button.map(|x| x.clicked.subscribe(|_| MyMsg::OkClicked))}
                {self.cancel_button.mount().map(MyMsg::CancelButton)}
                {self.cancel_button.map(|x| x.clicked.subscribe(|_| MyMsg::CancelClicked))}
            </>
        ).into()
    }
}

impl App for MyApp {
    fn update(&mut self, msg: Self::Message, ctx: Context<Self::Message>) -> Updated {
        match msg {
            MyMsg::OkButton(msg) => {
                self.ok_button.update(msg, ctx.map(MyMsg::OkButton))
            }
            MyMsg::CancelButton(msg) => {
                self.cancel_button.update(msg, ctx.map(MyMsg::CancelButton))
            }
            MyMsg::OkClicked => {
                println!("Ok Clicked!");
                Updated::no()
            }
            MyMsg::CancelClicked => {
                println!("Cancel Clicked!");
                Updated::no()
            }
        }
    }
}

Structs

Component
A Component wraps a Render or App type. A Component is mounted into a DOM and is tracked by the Runtime. This allows the runtime to call render() only if re-render is requird. This avoids re-rendering the whole DOM in each cycle and reduces the size of the DOM to be diffed.

Updated
Allows the update() cycle of an application or component to signal the runtime what portion of the DOM requires re-rendering.