Struct greenhorn::component::Component

pub struct Component<T: Render + Send> { /* fields omitted */ }

A Component wraps a Render or App type. A Component is mounted into a DOM and is tracked by the Runtime. This allows the runtime to call render() only if re-render is requird. This avoids re-rendering the whole DOM in each cycle and reduces the size of the DOM to be diffed.

Components can be mounted to a DOM using the Component::mount() function. Each component has an assigned id() which allows it to be identified by the runtime. Components that require re-rendering are recorded with their id() in the Updated type.

The underlying data type can be accessed using the lock(), map() or transmute() functions.

For a more detailed example, refer to the module level documentation.

Implementations

impl<T: 'static + Render + Send> Component<T>

pub fn new(inner: T) -> Self

Creates a new Component by taking ownership of the underlying type.

pub fn lock(&self) -> MutexGuard<'_, T>

Locks the underlying data for modification and returns a reference to it

pub fn id(&self) -> Id

Returns the unique Id associated with this component

The id() is used to identify this component in the DOM hierarchy. The executing runtime understands, based on the result of the update() cycle, which components require re-rendering.

pub fn render(&self) -> Node<T::Message>

Calls the render() function of the underlying Render object.

pub fn map<R, F: Fn(&T) -> R>(&self, fun: F) -> R

Applies a function to the underlying data and returns it’s result. The function receives an immutable reference.

pub fn transmute<R, F: FnOnce(&mut T) -> R>(&self, fun: F) -> R

Applies a function to the underlying data and returns it’s result. The function receives an mutable reference.

pub fn mount(&self) -> Node<T::Message>

Mounts the component to the DOM.

Usually, this function in chained with a .map() call on the result to map it to the Node with the appropriate message type.

Example

struct MyApp {
    ok_button: Component<Button>,
}

enum MyMsg {
    OkButton(ButtonMsg)
}

impl Render for MyApp {
    type Message = MyMsg;

    fn render(&self) -> Node<MyMsg> {
        html!(
            <div>
                {self.ok_button.mount().map(MyMsg::OkButton)}
            </>
        ).into()
    }
}

impl<T: 'static + App + Send> Component<T>

If the underlying type is App the Component also provides access to the App::update() and the App::mount() functions.

pub fn update(&mut self, msg: T::Message, ctx: Context<T::Message>) -> Updated

Call the update() function of the underlying object

pub fn on_mount(&mut self, ctx: Context<T::Message>)

Call the mount() function of the underlying object