Module greenhorn::context

The context module exports the Context type, which is passed into each update() cycle and allows interfacing with system services.

Context objects support the following features:

Emitting events - This may trigger a new update() cycle if an Event was subscribed to.

Loading CSS or JS on the frontend

Spawning futures or streams - the results of either trigger a new update() cycle.

Propagating or invoking DOM events on the frontend

Showing system dialogs

Quitting the application

For more details, refer to the Context type.

Structs

Context
Context objects are passed into the update() function of the application and allow interacting with the component hierarchy, controlling the application lifecycle and provide access to system services.

EventPropagate