Struct greenhorn::context::Context

pub struct Context<T: 'static + Send> { /* fields omitted */ }

Context objects are passed into the update() function of the application and allow interacting with the component hierarchy, controlling the application lifecycle and provide access to system services.

Methods

impl<T: Send + 'static> Context<T>

pub fn emit<D: Any>(&self, event: &Event<D>, data: D)

Emits an event from the current component.

This will result in an update() call for all components subscribed to this event. Emitting Events triggers new update() cycles of the application and thus allows interchanging messages between components.

pub fn load_css<Css: Into<String>>(&self, css: Css)

Loads a CSS string on the frontend

pub fn run_js<Js: Into<String>>(&self, js: Js)

Runs a piece of javascript code on the frontend

pub fn run_service<S, F>(&self, service: S, fun: F) where
S: 'static + Service,
F: 'static + Send + Fn(S::Data) -> T,

Spawn a Service using a mapping function to map the services data items to the current message type

pub fn spawn<Fut: 'static + Send + Future<Output = T>>(&self, fut: Fut)

Spawns a future. The result of the future will be used to update() the application.

pub fn spawn_blocking<Fut: 'static + Send + Future<Output = T>>(&self, fut: Fut)

Spawns a future which contains blocking operations. This future might be spawned on a different thread-pool to avoid stalling non-blocking futures.

pub fn subscribe<S: 'static + Send + Stream<Item = T>>(&self, stream: S)

Subscribe to a stream. Each item the stream issues will be used to udpate() the application.

pub fn map<U: Send + 'static, F: 'static + Send + Sync + Fn(U) -> T>(
&self,
fun: F
) -> Context<U>

Maps this context object to a new a new message type

pub fn propagate(&self, e: DomEvent)

Propagates a previously intercepted DOM event to the frontend.

pub fn default_action(&self, e: DomEvent)

Propagates a DOM event to the frontend executing the default action.

This is useful in case a DOM event handler was registered with prevent_default(). and the application has determined that the default action should be executed anyway.

pub fn propagate_and_default(&self, e: DomEvent)

Propagates a previously intercepted js event to the frontend and execute the default action.

pub fn dialog<D: 'static + Dialog, F: 'static + Fn(D::Msg) -> T>(
&self,
dialog: D,
fun: F
)

Opens a dialog on the frontend.

Once the dialog resolves, the fun function maps the dialog message type to the message type of the application. Refer to the dialog module for builtin dialogs. Note that only one dialog is shown at the same time, since dialogs are usully modal. However, multiple dialogs may be enqueued, which will then be shown one after the next.

pub fn quit(&self)

Quits the currently running application and shuts down all associated services.