Module greenhorn::dialog

Allows spawning native dialogs such as file-open dialogs and retrieving their results. This module implements a set of built-in system dialogs which can be shown using Context::dialog().

This module supports spawning:

FileSaveDialog

FileOpenDialog

MessageBox

This module just provides a common interface for these dialogs. However, the dialogs actually need to be implemented by a frontend implementation. This library provides the feature flag native-dialog which makes use of the tinyfiledialogs-rs library to show native system dialogs.

Structs

FileFilter
Represents a filter in a FileOpenDialog or a FileSaveDialog. A filter consists of a name and a list of file extensions (such as "js" or "rs").

FileOpenDialog
Represents a system dialog for opening one or multiple files. Once the dialog is closed it resolves to a FileOpenMsg.

FileSaveDialog
Represents a system dialog for saving a file. Once the dialog is closed it resolves to a FileSaveMsg.

MessageBox
Represents the message box dialog.

Enums

FileOpenMsg
Message emitted when closing a FileOpenDialog.

FileSaveMsg
Message type a FileSaveDialog resolves to when closed.

MessageBoxResult
The result of the message box the dialog resolves to once the dialog has been closed.

MsgBoxIcon
Defines the message box icon

MsgBoxType
Defines the type of message box according to the buttons on the window

Traits

Dialog
Interface for modal dialogs. A dialog may be spawned using the Context::dialog() function. Once a dialog closes it resolves to a result captured using the Msg type.