Trait greenhorn::dialog::Dialog

pub trait Dialog: Sealed + Serialize + DeserializeOwned + Sized {
    type Msg: DeserializeOwned;
    fn type_name() -> &'static str;

    fn resolve(self, data: JsonValue) -> Result<Self::Msg, Error> { ... }

    fn serialize(&self) -> JsonValue { ... }
}

Interface for modal dialogs. A dialog may be spawned using the Mailbox::dialog() function. Once a dialog closes it resolves to a result captured using the Msg type.

Associated Types

type Msg: DeserializeOwned

Loading content...
Required methods

fn type_name() -> &'static str

Must return a type name uniquely identifying this type within all Dialog types. This allows the resulting json to be associated to a type upon deserialization.

Loading content...
Provided methods

fn resolve(self, data: JsonValue) -> Result<Self::Msg, Error>

Called by the runtime to produce a result based on the received data from the dialog after the user has closed it.

fn serialize(&self) -> JsonValue

Serializes the current object into a json string. Also inserts a __type__ field.

Loading content...