Module greenhorn::dom

This module exposes the DomEvent data type which is used to communicate DOM events from the frontend to the backend.

DomEvents are the principal data type for frontend to backend communication. When subscribing to a event on the DOM a handler function is registered in the application runtime, which maps a DomEvent to a the message type of the running [../trait.App.html].

When the subscribed event is triggered on the frontend, a DomEvent is created, mapped with the handler function and subsequently passed into the update() cycle of the application.

Structs

BaseEvent
Minimal data type to represent unsupported HTML Events.

KeyboardEvent
Mapping of the HTML KeyboardEvent

ModifierState
Defines whether a modifier is currently pressed

MouseEvent
Mapping of the HTML MouseEvent

Point
Maps to an (x,y) coordinate tuple for HTML MouseEvents

WheelEvent
Mapping of the HTML WheelEvent

Enums

DomEvent
The DomEvent enum maps HTML Events into a rust datatype.

InputValue
Maps to the value attribute of HTMLElement.