Module greenhorn::event

This module defines the Event type which facilitates communication between components.

During the render() phase, it is possible to subscribe to Event objects. Event subscriptions are part of the frontend state and are represented in Node objects. In case an Event is .emit()-ed during the update() phase the respective event subscriptions are queried and trigger a new update() cycle.

Oftentimes, components may want to inform other components about a change of their state. In such a case it may expose Event objects, to which other components can subscribe to.

Note that events can only be subsribed to once, since the data of the event emission is moved into the update() cycle of the application.

Example

struct CheckBox {
    change: Event<bool>,
    checked: bool,
}

enum CheckBoxMsg {
    Click(DomEvent), KeyDown(DomEvent)
}

impl App for CheckBox {
    fn update(&mut self, msg: Self::Message, ctx: Context<Self::Message>) -> Updated {
        match msg {
            CheckBoxMsg::Click(evt) => {
                self.checked = !self.checked;
                ctx.emit(&self.change, self.checked);
                Updated::yes()
            }
            CheckBoxMsg::KeyDown(evt) => {
                let key = evt.into_keyboard().unwrap().key;
                if key == "Enter" || key == "Space" {
                    self.checked = !self.checked;
                    ctx.emit(&self.change, self.checked);
                    Updated::yes()
                } else {
                    Updated::no()
                }
            }
        }
    }
}

impl Render for CheckBox {
    type Message = CheckBoxMsg;

    fn render(&self) -> Node<Self::Message> {
        // render a fancy checkbox and subscribe to DOM events...
    }

}

struct Parent {
    checkbox: CheckBox
}

enum ParentMsg {
    // ...
    CheckBox(CheckBoxMsg),
    CheckedChanged(bool),
}

impl Render for Parent {
    type Message = ParentMsg;

    fn render(&self) -> Node<Self::Message> {
        html!(<div>
                {self.checkbox.render().map(ParentMsg::CheckBox)}
                {self.checkbox.change.subscribe(ParentMsg::CheckedChanged)}
            </div>
        ).into()
    }
}

Structs

Emission
This type is created if a Event is emit()-ed.

Event
Events allow connecting different components with each other.

Subscription
Represents the result of an Event.subscribe() call.