Struct greenhorn::node::Node

pub struct Node<T: 'static + Send>(_);

Represents a DOM node which might emit a message of type T.

Node objects are produced by Render::render() functions. Each Render implementation has an associated Message type, which is emitted by the returned Nodes, thus feeding information back into the update() cycle of the application.

Nodes may be constructed using a NodeBuilder:

Elements by using an ElementBuilder

Text

Component instances

Event subscriptions

Blobs

Furthermore, Nodes can be constructed using the html!() and svg!() macros. Nodes may not necessarily lead to a node rendered in the DOM but merely represent state in the frontend.

Example

struct MyRenderable {
    blob: Blob,
    button: Component<Button>,
};

impl Render for MyRenderable {
    type Message = MyMsg;

    fn render(&self) -> Node<Self::Message> {
        Node::html().elem("div").class("primary").id("my-div")  // an HTML element
            .add(Node::text("Some Text"))                       // a text node
            .add(&self.blob)                                    // a blob
            .add(self.button.mount().map(MyMsg::ButtonMsg))     // a component
            .add(self.button.lock().clicked.subscribe(|_| MyMsg::Clicked)) // an event
            .build()
    }
}

Implementations

impl<T: 'static + Send> Node<T>

pub fn new_from_iter<S: Iterator<Item = Node<T>>>(nodes: S) -> Self

pub fn html() -> NodeBuilder<T>

Create a NodeBuilder for HTML elements

pub fn svg() -> NodeBuilder<T>

Create a NodeBuilder for SVG elements

pub fn text<S: ToString>(data: S) -> Self

Produce a text node

pub fn map<U: 'static + Send, F: 'static + Send + Fn(T) -> U>(
self,
fun: F
) -> Node<U>

Maps the message type of the node to a new message type

When nesting different Render implementations or components, the message types need to be nested as well. The Node::map() function allows the emitted messages of type T to be mapped to another message type U.

Example

enum ButtonMsg {
    KeyDown(DomEvent),
    Clicked(DomEvent),
}

enum MyAppMsg {
    ButtonOk(ButtonMsg),
    ButtonCancel(ButtonMsg),
}

fn render_button(text: &str) -> Node<ButtonMsg> {
    // ...
}

fn render() -> Node<MyAppMsg> {
    html!(
        <div>
            {render_button("Ok").map(MyAppMsg::ButtonOk)}
            {render_button("Cancel").map(MyAppMsg::ButtonCancel)}
        </>
    ).into()
}

pub fn map_shared<U: 'static + Send>(
self,
fun: Arc<Mutex<dyn Send + Fn(T) -> U + 'static>>
) -> Node<U>

same as Node::map() but uses a shared reference of an already created mapping function

pub fn empty_map<U: 'static + Send>(self) -> Node<U>

Maps Node() objects without providing a mapping-functions.

Panics in case there are listeners installed on this node or any child node. This allows mapping node-hierarchies without listeners efficiently without keeping the target message type around, for example when caching rendered nodes.

pub fn try_clone(&self) -> Option<Self>

Attempt to clone this Node.

If the Node has been mapped to a different type, the Node cannot be cloned anymore. Also, mounted components and event subscriptions cannot be cloned.