Module greenhorn::pipe

This modules defines an interface for the communication interface between frontend and backend.

The central trait is the Pipe type, which exposes both a Sender and a Receiver.

A Sender can be used to send TxMsg message, i.e. message from backend to the frontend. A Receiver can be used to receive RxMsg message, i.e. message from frontend to the backend.

Enums

RxMsg
Serializable message type to be sent from the frontend to the backend

TxMsg
Serializable message type to be sent from the backend to the frontend

Traits

Pipe
This trait defines the interface for sending and receiving messages from the Runtime.

Receiver
Receiver trait for receiving RxMsg objects

Sender
Sender trait for sending TxMsg objects