Trait greenhorn::pipe::Pipe

pub trait Pipe {
    type Sender: Sender;
    type Receiver: Receiver;
    fn split(self) -> (Self::Sender, Self::Receiver);
}

This trait defines the interface for sending and receiving messages from the Runtime.

Associated Types

type Sender: Sender

type Receiver: Receiver

Loading content...
Required methods

fn split(self) -> (Self::Sender, Self::Receiver)

Consumes this object and split into into a sender and receiver part.

Either of them may be sent to different threads and used independently.

Loading content...