Module greenhorn::service

Supports spawning tasks running on the frontend. Experimental, might be removed in the future. This module exposes the Service trait, which allows writing custom agents communicating with the frontend.

Structs

Mailbox
The mailbox type allows a Service to communicate with the frontend via custom messages.

Enums

RxServiceMessage
RxServiceMessage objects are sent by the frontend and received by the backend.

TxServiceMessage
A backend service may send TxServiceMessage objects to the frontend.

Traits

Service
The Service trait allows a type to act as an agent, which communicates over custom messages with the frontend and injects messages back into the update() cycle of the application.