Trait greenhorn::service::Service

pub trait Service {
    type Data: 'static + Send;
    type DataStream: Stream<Item = Self::Data> + Send;
    fn start(self, mailbox: Mailbox) -> Self::DataStream;
}

The Service trait allows a type to act as an agent, which communicates over custom messages with the frontend and injects messages back into the update() cycle of the application.

After the start() method has been called, a service produces Data items which are injected back into the update() cycle of the application. The service itself can send and receive messages from the frontend using the Mailbox object it has obtained from the invocation of start().

Example

use futures::{channel::mpsc, StreamExt};
use greenhorn::platform;

struct MyService {}

impl Service for MyService {
   type Data = i32;
   type DataStream = mpsc::UnboundedReceiver<i32>;

   fn start(self, mut mailbox: Mailbox) -> Self::DataStream {
       mailbox.run_js("ctx.send('123')");
       let (tx, rx) = mpsc::unbounded();
       platform::spawn(async move {
           loop {
               if let Some(RxServiceMessage::Frontend(data)) = mailbox.next().await {
                   // this will receive '123' from the frontend
                   let data = i32::from_str(&data).unwrap();
                   // send the i32 into the update() cycle of the application
                   tx.unbounded_send(data).unwrap();
               } else {
                   break;
               }
           }
       });
       rx
   }
}

Associated Types

type Data: 'static + Send

The data item the service emits.

These items are wrapped in messages and passed to the update() cycle of the application

type DataStream: Stream<Item = Self::Data> + Send

A stream which produces the services data items

Loading content...
Required methods

fn start(self, mailbox: Mailbox) -> Self::DataStream

Starts a service and consumes it.

The returned DataStream may produce items to be passed into the update() loop of the applications. The service should subscribe to messages from the Mailbox (a Stream).

Loading content...