Trait greenhorn::App

pub trait App: Render {
    fn update(
        &mut self, 
        msg: Self::Message, 
        ctx: Context<Self::Message>
    ) -> Updated;

    fn mount(&mut self, _ctx: Context<Self::Message>) { ... }
}

Marks a type as render-able as well as update-able

The update method should modify the state of the object and return an Updated object marking whether the update should requires a re-render of the DOM.

At the same time an optional mount() function may be provided to allow a component to hook into the startup cycle of a component or application.

Example

use greenhorn::prelude::*;
use greenhorn::html;

struct MyApp {
    text: String,
}

enum MyMsg {
    Clicked(DomEvent),
    KeyDown(DomEvent),
}

impl Render for MyApp {
    type Message = MyMsg;

    fn render(&self) -> Node<Self::Message> {
        html!(
            <div #my-app>
                <button type="button"
                        @keydown={MyMsg::KeyDown}
                        @mousedown={MyMsg::Clicked}>
                    {&self.text}
                </>
            </>
        ).into()
    }
}

impl App for MyApp {
    fn update(&mut self,msg: Self::Message,ctx: Context<Self::Message>) -> Updated {
        match msg {
            MyMsg::Clicked(evt) => self.text = "Button clicked!".into(),
            MyMsg::KeyDown(evt) => self.text = "Button keypress!".into()
        }
        Updated::yes()
    }

}

Required methods

fn update(&mut self, msg: Self::Message, ctx: Context<Self::Message>) -> Updated

Modify the state of this object based on the received Message. Returns an Updated which should mark whether this component is to be re-rendered.

Each component is required to dispatch messages to its child components if applicable.

Loading content...
Provided methods

fn mount(&mut self, _ctx: Context<Self::Message>)

Shall be called upon application startup. A parent component is required to call this function of all child components.

Loading content...