Trait greenhorn::Render

pub trait Render {
    type Message: 'static + Send;
    fn render(&self) -> Node<Self::Message>;
}

Marks a type as render-able to the DOM.

A Render type also defines a Message type, which defines a type being emitted by the Nodes it emits. Implementing this trait is the minimum requirement to create a Component.

Example

struct Button {
    text: String,
}

enum ButtonMessage {
    Clicked(DomEvent),
    KeyDown(DomEvent),
}

impl Render for Button {
    type Message = ButtonMessage;

    fn render(&self) -> Node<Self::Message> {
        html!(
            <button type="button"
                    @keydown={ButtonMessage::KeyDown}
                    @mousedown={ButtonMessage::Clicked}>
                {&self.text}
            </>
        ).into()
    }
}

Associated Types

type Message: 'static + Send

Defines a type which is emitted when capturing DOM events or component events

Typically this is an enum.

Loading content...
Required methods

fn render(&self) -> Node<Self::Message>

Renders this type to a DOM.

The nodes may emit messages of type Self::Message which are passed again into the update() cycle of the application.

Loading content...