Struct greenhorn::websockets::WebSocketPipe

pub struct WebSocketPipe { /* fields omitted */ }

The WebSocketPipe type implements a Pipe on top of WebSockets.

The WebSocketPipe acts as a async server only accepting a single connection. Once created, the WebSocketPipe can be .split() using the method on the Pipe trait. The resulting WebSocketSender and WebSocketReceiver structs may be used to send and receive messages over the resulting websocket connection. The respective Stream and Sink implementations are closed once the connection has been closed.

Example

let pipe = WebSocketPipe::listen_to_addr(SocketAddr::from_str("127.0.0.1:0").unwrap());
let (sender, mut receiver) = pipe.split();
task::spawn(async move {
    while let Some(_msg) = receiver.next().await {
        // do something with msg...
        // and send something
        // sender.send(....).await;
    }
});

Methods

impl WebSocketPipe

pub fn listen_to_addr(addr: SocketAddr) -> WebSocketPipe

Starts listening to a given SocketAddr

pub fn listen_to_socket(listener: TcpListener) -> WebSocketPipe

Start listening to TcpListener.

pub fn local_addr(&self) -> SocketAddr

Returns the local address the WebSocket server is listening on.

pub fn port(&self) -> u16

Returns the port the WebSocket server is listening on

This is especially useful when using the OS to automatically assign a port to your connection (by listening to port ":0")